+++ src/rules/media-query-list-comma-newline-after/__tests__/index.js
@@ -19,9 +19,14 @@ testRule(rule, {
   }, {
     code: "@media screen and (color),\nprojection and (color) {}",
   }, {
+    code: "@media screen and (color),\n\nprojection and (color) {}",
+  }, {
     code: "@media screen and (color) ,\n  projection and (color) {}",
   }, {
     code: "@media screen and (color) ,\r\n  projection and (color) {}",
+    description: "CRLF",
+  }, {
+    code: "@media screen and (color) ,\r\n\r\n  projection and (color) {}",
     description: "CRLF",
   }, {
     code: "@media screen and (color)\n,\n\t\t\tprojection and (color) {}",